requirements.txt
streamlit>=1.32.0
groq>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit-mic-recorder>=0.0.8
pandas>=2.0.0
python-dateutil>=2.8.0
elevenlabs>=1.0.0